backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "uvicorn",
    "requests",
    "beautifulsoup4",
    "google-generativeai",
    "python-dotenv",
    "playwright",
    "cssutils",
]
